+++ io-service/src/main/java/com/example/ioservice/dto/OrderedProductDto.java
@@ -1,0 +1,4 @@
+package com.example.ioservice.dto;
+
+public record OrderedProductDto(Long productId, String productName, Float price, Long quantity) {
+}