+++ io-service/src/main/java/com/example/ioservice/dto/ProductDto.java
@@ -1,4 +1,4 @@
 package com.example.ioservice.dto;
 
-public record ProductDto(Long productId, String productName, Float price, String description, Long ownerId, Integer quantity) {
+public record ProductDto(Long productId, String productName, Float price, String description, Long sellerId, Integer quantity) {
 }