+++ io-service/src/main/java/com/example/ioservice/dto/UserDto.java
@@ -1,4 +1,4 @@
 package com.example.ioservice.dto;
 
-public record UserDto(Long userId, String username, String email, String address) {
+public record UserDto(Long userId, String address, String phoneNumber) {
 }